refactor: add folders for project structure
Split classes in two groups of controls and table with futher
subdivision into folders

Change import pathes
