+++ index.html
@@ -58,6 +58,6 @@
     <div class="table-wrapper"></div>
 
 </body>
-<script src="main.js" type="module"></script>
+<script src="js/main.js" type="module"></script>
 
 </html>
