+++ js/main.js
@@ -1,5 +1,5 @@
-import { Table } from './Table.js';
-import { InteractiveTable } from './InteractiveTable.js';
+import { Table } from './table/Table.js';
+import { InteractiveTable } from './table/InteractiveTable.js';
 
 const table = new Table(4, 4, document.querySelector('.table-wrapper'));
